adding in the olm config file so that an operator bundle is generated (#7)
Issue #, if available:
- Fixes: aws-controllers-k8s/community#1305

Description of changes:
Adding in the olmconfig.yaml so that an operator bundle can be generated for OperatorHub. I think I got the branding correct for this controller, with one exception noted below. Since the service package name is different then the branded name, the keywords will appear like the below in the final `CSV` that is generated:

```
keywords:
  - emrcontainers
  - aws
  - amazon
  - ack
 ```
This comes from a template file in the code gen, and I don't think it makes sense to try to do an override for this one controller, but wanted to point this out incase there are concerns.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

Signed-off-by: Adam D. Cornett <adc@redhat.com>